+++ CHANGES.rst
@@ -10,6 +10,9 @@ Change Log
 
 * ``ralph_device_id`` get automatically cleaned when when Device linked to it gets deleted.
 
+* Added partial and exact searches to assets.
+
+
 
 1.2.6
 ~~~~~
@@ -23,6 +26,7 @@ Released on August 08, 2013
 * Added link to the Pricing App.
 
 * Added field: last modification, asset_id to csv file.
+
 
 
 1.0.0